+++ Source/Fluxor/DependencyInjection/DependencyScanners/MiddlewareClassesDiscovery.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
@@ -6,13 +8,26 @@ namespace Fluxor.DependencyInjection.DependencyScanners
 {
 	internal static class MiddlewareClassesDiscovery
 	{
-		internal static AssemblyScanSettings[] FindMiddlewareLocations(IEnumerable<Assembly> assembliesToScan)
+		internal static DiscoveredMiddleware[] FindMiddlewares(
+			IServiceCollection services,
+			IEnumerable<Assembly> assembliesToScan,
+			IEnumerable<Type> manuallyIncludedMiddlewares)
 		{
-			return assembliesToScan
-				.SelectMany(x => x.GetTypes().Where(t => t.GetInterfaces().Any(i => i == typeof(IMiddleware))))
-				.Select(x => new AssemblyScanSettings(x.Assembly, x.Namespace))
-				.Distinct()
+			var manuallyIncludedMiddlewaresLookup = new HashSet<Type>(manuallyIncludedMiddlewares);
+			DiscoveredMiddleware[] discoveredMiddlewares = assembliesToScan
+				.SelectMany(a => a.GetTypes().Where(t => t.GetInterfaces().Any(i => i == typeof(IMiddleware))))
+				.Select(t =>
+					new DiscoveredMiddleware(
+						implementingType: t,
+						autoLoaded: 
+							manuallyIncludedMiddlewares.Contains(t)
+							|| t.GetCustomAttribute(typeof(AutomaticallyIncludeMiddlewareAttribute)) != null))
 				.ToArray();
+
+			foreach (DiscoveredMiddleware discoveredMiddleware in discoveredMiddlewares)
+				if (discoveredMiddleware.AutoLoaded)
+					services.AddScoped(discoveredMiddleware.ImplementingType);
+			return discoveredMiddlewares;
 		}
 	}
 }