+++ Source/Fluxor/DependencyInjection/ServiceCollectionExtensions.cs
@@ -29,16 +29,12 @@ namespace Fluxor
 			var options = new Options(serviceCollection);
 			configure?.Invoke(options);
 
-			// Register all middleware types with dependency injection
-			foreach (Type middlewareType in options.MiddlewareTypes)
-				serviceCollection.AddScoped(middlewareType);
-
 			IEnumerable<AssemblyScanSettings> scanIncludeList = options.MiddlewareTypes
 				.Select(t => new AssemblyScanSettings(t.Assembly, t.Namespace));
 
 			DependencyScanner.Scan(
 				options: options,
-				serviceCollection: serviceCollection,
+				services: serviceCollection,
 				assembliesToScan: options.AssembliesToScan,
 				scanIncludeList: scanIncludeList);
 			serviceCollection.AddScoped(typeof(IState<>), typeof(State<>));