+++ Tests/Fluxor.UnitTests/DependencyInjectionTests/MiddlewareDiscoveryTests/AutoMiddleware.cs
@@ -1,0 +1,7 @@
+﻿namespace Fluxor.UnitTests.DependencyInjectionTests.MiddlewareDiscoveryTests
+{
+	[AutomaticallyIncludeMiddleware]
+	public class AutoMiddleware : Middleware
+	{
+	}
+}